requirements.txt
PyYaml
PyGithub